refactor: move processFeedingCows to its own file
