+++ src/game-logic/reducers/index.test.js
@@ -1,5 +1,4 @@
 import { shapeOf, testCrop, testItem } from '../../test-utils'
-import { OUT_OF_COW_FEED_NOTIFICATION } from '../../strings'
 import {
   ACHIEVEMENT_COMPLETED,
   COW_ATTRITION_MESSAGE,
@@ -11,14 +10,11 @@ import {
   PRICE_SURGE,
 } from '../../templates'
 import {
-  COW_FEED_ITEM_ID,
   COW_GESTATION_PERIOD_DAYS,
   COW_HUG_BENEFIT,
   COW_MILK_RATE_SLOWEST,
   COW_FERTILIZER_PRODUCTION_RATE_SLOWEST,
-  COW_WEIGHT_MULTIPLIER_MAXIMUM,
   COW_WEIGHT_MULTIPLIER_MINIMUM,
-  COW_WEIGHT_MULTIPLIER_FEED_BENEFIT,
   MAX_ANIMAL_NAME_LENGTH,
   MAX_DAILY_COW_HUG_BENEFITS,
   MAX_LATEST_PEER_MESSAGES,
@@ -404,122 +400,6 @@ describe('computeStateForNextDay', () => {
     expect(firstRow[0].daysWatered).toBe(1)
     expect(firstRow[0].daysOld).toBe(1)
     expect(todaysNotifications).toBeEmpty()
-  })
-})
-
-describe('processFeedingCows', () => {
-  let state
-
-  beforeEach(() => {
-    state = {
-      cowInventory: [],
-      inventory: [],
-    }
-  })
-
-  describe('player has no cow feed', () => {
-    beforeEach(() => {
-      state.cowInventory = [generateCow({ weightMultiplier: 1 })]
-      state.newDayNotifications = []
-    })
-
-    test('cows weight goes down', () => {
-      const {
-        cowInventory: [{ weightMultiplier }],
-        newDayNotifications,
-      } = fn.processFeedingCows(state)
-
-      expect(weightMultiplier).toEqual(1 - COW_WEIGHT_MULTIPLIER_FEED_BENEFIT)
-      expect(newDayNotifications).toEqual([
-        {
-          message: OUT_OF_COW_FEED_NOTIFICATION,
-          severity: 'error',
-        },
-      ])
-    })
-  })
-
-  describe('player has cow feed', () => {
-    beforeEach(() => {
-      state.cowInventory = [
-        generateCow({ weightMultiplier: 1 }),
-        generateCow({ weightMultiplier: 1 }),
-      ]
-      state.newDayNotifications = []
-    })
-
-    describe('there are more feed units than cows to feed', () => {
-      test('units are distributed to cows', () => {
-        state.inventory = [{ id: COW_FEED_ITEM_ID, quantity: 4 }]
-        const {
-          cowInventory,
-          inventory: [{ quantity }],
-          newDayNotifications,
-        } = fn.processFeedingCows(state)
-
-        expect(cowInventory[0].weightMultiplier).toEqual(
-          1 + COW_WEIGHT_MULTIPLIER_FEED_BENEFIT
-        )
-        expect(cowInventory[1].weightMultiplier).toEqual(
-          1 + COW_WEIGHT_MULTIPLIER_FEED_BENEFIT
-        )
-        expect(quantity).toEqual(2)
-        expect(newDayNotifications).toEqual([])
-      })
-    })
-
-    describe('there are more cows to feed than feed units', () => {
-      test('units are distributed to cows and remainder goes hungry', () => {
-        state.inventory = [{ id: COW_FEED_ITEM_ID, quantity: 1 }]
-        const {
-          cowInventory,
-          inventory,
-          newDayNotifications,
-        } = fn.processFeedingCows(state)
-
-        expect(cowInventory[0].weightMultiplier).toEqual(
-          1 + COW_WEIGHT_MULTIPLIER_FEED_BENEFIT
-        )
-        expect(cowInventory[1].weightMultiplier).toEqual(
-          1 - COW_WEIGHT_MULTIPLIER_FEED_BENEFIT
-        )
-        expect(inventory).toHaveLength(0)
-        expect(newDayNotifications).toEqual([
-          {
-            message: OUT_OF_COW_FEED_NOTIFICATION,
-            severity: 'error',
-          },
-        ])
-      })
-    })
-
-    describe('mixed set of weightMultipliers with unsufficient cow feed units', () => {
-      test('units are distributed to cows and remainder goes hungry', () => {
-        state.cowInventory = [
-          generateCow({ weightMultiplier: COW_WEIGHT_MULTIPLIER_MAXIMUM }),
-          generateCow({ weightMultiplier: COW_WEIGHT_MULTIPLIER_MAXIMUM }),
-          generateCow({ weightMultiplier: 1 }),
-          generateCow({ weightMultiplier: 1 }),
-        ]
-        state.inventory = [{ id: COW_FEED_ITEM_ID, quantity: 3 }]
-
-        const { cowInventory, inventory } = fn.processFeedingCows(state)
-
-        expect(cowInventory[0].weightMultiplier).toEqual(
-          COW_WEIGHT_MULTIPLIER_MAXIMUM
-        )
-        expect(cowInventory[1].weightMultiplier).toEqual(
-          COW_WEIGHT_MULTIPLIER_MAXIMUM
-        )
-        expect(cowInventory[2].weightMultiplier).toEqual(
-          1 + COW_WEIGHT_MULTIPLIER_FEED_BENEFIT
-        )
-        expect(cowInventory[3].weightMultiplier).toEqual(
-          1 - COW_WEIGHT_MULTIPLIER_FEED_BENEFIT
-        )
-        expect(inventory).toHaveLength(0)
-      })
-    })
   })
 })
